+++ src/ui/compendium/plugin.rs
@@ -5,7 +5,7 @@ use crate::game::input::messages::MouseClicked;
 use crate::state::{MenuState, MetaGameState, PauseMenuState};
 use crate::ui::plugin::ButtonActionSet;
 
-use super::components::{BackButton, CompendiumState, ScrollableCompendiumContainer};
+use super::components::{BackButton, CompendiumState, DetailPanel, ScrollableCompendiumContainer};
 use crate::ui::systems::{escape_to_landing, escape_to_pause_main, handle_scroll};
 
 use super::systems;
@@ -39,6 +39,7 @@ impl Plugin for MainMenuCompendiumPlugin {
                 Update,
                 (
                     handle_scroll::<ScrollableCompendiumContainer>,
+                    handle_scroll::<DetailPanel>,
                     escape_to_landing,
                     systems::rebuild_on_state_change,
                 )
@@ -79,6 +80,7 @@ impl Plugin for PauseMenuCompendiumPlugin {
             Update,
             (
                 handle_scroll::<ScrollableCompendiumContainer>,
+                handle_scroll::<DetailPanel>,
                 escape_to_pause_main,
                 systems::rebuild_on_state_change,
             )
@@ -116,6 +118,7 @@ impl Plugin for MetaGameCompendiumPlugin {
                 Update,
                 (
                     handle_scroll::<ScrollableCompendiumContainer>,
+                    handle_scroll::<DetailPanel>,
                     escape_to_wizard_tower,
                     systems::rebuild_on_state_change,
                 )